requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==4.2.15
django-colorfield==0.11.0
django-filter==24.3
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-xml==2.0.0
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
PyJWT==2.9.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
psycopg2-binary==2.9.6